Virtual Wardrobe Assistant/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
python-multipart==0.0.6
Flask-Migrate==4.0.5
opencv-python==4.8.1.78
werkzeug==2.3.7
wtforms==3.1.1
streamlit==1.44.1 